Modularise viewer
