+++ templates/_layouts/presenter.html
@@ -4,8 +4,6 @@
 {% block layout %}
 {% assets filters="jsmin", output="dist/js/presenter.min.js",
     "../node_modules/tether-shepherd/dist/js/shepherd.js",
-    "../node_modules/openseadragon/build/openseadragon/openseadragon.js",
-    "../node_modules/openseadragonselection/dist/openseadragonselection.js",
     "../node_modules/imagesloaded/imagesloaded.pkgd.js"
     %}
     <script type="text/javascript" src="{{ ASSET_URL }}"></script>
